Remove pillow exact dependency (#283)
That exact dependency was added in #261 at Jun 19. It's unclear to me what version conflicted because it appeared to be the latest released version at that date ( https://pypi.org/project/Pillow/#history ).

The problem is that `scikit-image==0.14.0` requires `pillow [required: >=4.3.0]` (got with `pipenv graph`). So this rule makes it impossible to install the requirements: `Could not find a version that matches Pillow==4.1.1,>=4.1.1,>=4.3.0`.

I started running `make docs` but after 15min it didn't finish the first build so I don't really know how to reproduce the initial error.