Add custom dataset and dataloader tutorial for C++
I was struggling with creating a custom dataset and dataloader for the
C++ frontend partially because I couldn't find a good minimal example
and ran into a few gotchas when trying to model something after the
MNIST dataset.  This tutorial might help others avoid the pitfalls that
I ran into.

Signed-off-by: David Pollack <david@da3.net>