[TorchScript] Fix Intro to TorchScript tutorial
Summary:
There is a section of the Intro to TorchScript tutorial that is supposed
to illustrate that tracing cannot capture control flow but scripting can.
However, the generated IR for both cases is the same because the tutorial
only prints out the IR for the outer module, not the inner module that
contains the control flow. This commit fixes this issue by printing the IR
for both the inner and outer module to clearly show that tracing does not
capture control flow, but scripting does.

Testing:
Ran the tutorial file.