Add loss_fn as an input to the test function (#1529)
* Add loss_fn as an input to the test function

The function `test()` depends on `loss_fn` just like `train()` does. It is better to explicitly provide `loss_fn` as an argument instead of relying on a global `test_fn` object. Also, this is more consistent with how `train()` is defined. And it also makes it more explicit that `train()` has a dependency on `optimizer`, while `test()` doesn't.

* Update quickstart_tutorial.py

Co-authored-by: Holly Sweeney <77758406+holly1238@users.noreply.github.com>