fix model(Remove unwanted relu layer.)
A relu layer was inserted just before the cross entropy error was calculated, and it seemed that the loss was not calculated correctly, so I removed the relu layer.
Accordingly, the structure of the NeuralNetwork class in buildmodel_tutorial.py was also changed.

FYI:
After fixing the model, we tried the task in the environment at hand, and the accuracy went from about 52% to about 71%.

If you have time, I hope you will take a look.