Missing sample when loading csv file (#1708)
When reading a file with the ``read_csv`` method, by default the first line of the file is assumed to contain header information. In the example file mentioned in the tutorial, the first line contains data and if the ``names`` attribute of the ``read_csv`` method is not informed, the first line will be considered as a header. This way, when loading the dataset the first sample will be lost.

Co-authored-by: Holly Sweeney <77758406+holly1238@users.noreply.github.com>