Redirect old `two_layer` tutorials to `polynomial`.
The old artifacts are still accessible in the legacy form with broken layout.

This commit redirects the old pages to the new pages and
remove related assets.

`tf_two_layer` was deleted in #1265 but I am not sure if it should be deleted,
so not changing it in this commit.