Fixing timing scheme in cpp_extension tutorial (#1851)
* Fixing timing scheme in cpp_extension tutorial

Print statement incorrectly stated `us` when it was in `100ms` units.
For clarity I changed it to `s` units and removed the multiplication.
This could have intended to be averaged but total time works just as
well.

* type -> device

`type()` is depreciated and it looks like `device()` is appropriate
here. There are other `type` commands below and the error message says
that `option` can be a drop-in but I'm unsure if this is appropriate for
those calls.