rename confusing variable name in LM tutorial (#1953)
The tutorial "Language Modeling with nn.Transformer and TorchText" contains
code snippets with variables named `batch_size`. The issue is that in some
places, `batch_size` means that number of sequences in a batch, and in other
places it means the number of tokens in each batch sequence. This inconsistency
was solved in this commit: `batch_size` was replaced with `seq_len` in the two
places where it has the latter meaning.

Co-authored-by: Svetlana Karslioglu <svekars@fb.com>