Pin importlib_metadata<5.0 in requirements.txt
v5.0.0 causes issues with python 3.7: https://github.com/python/importlib_metadata/issues/411

This was raised in the context of https://github.com/pytorch/tutorials/issues/2090

This PR pins the version to <5.0 to circumvent this. Maybe a better fix would be to bump the python version in CI to 3.8, but that's something to discuss more broadly.