Update translation_transformer.ipynb  env packages
Add some required packages and file download scripts to make the notebook runnable without problems. Before this change, it raised an error for the lack of data ` [E050] Can't find model 'de_core_news_sm.' It doesn't seem to be a Python package or a valid path to a data directory.` And raised the `torchdata` requirement error.