Use the correct batch values in the output
The current logic uses enumerate counter (i.e. variable `batch`). It displays the loss for
trained data in an increment of 100 using batch % 100 == 0. i.e. 1 batch, 101 batch, and so on
for the given dataset. This maps to batch =0, 100, and so on). So for the first batch, the loss
displayed should be [ 64/60000] instead of [ 0/60000]). For the second it should be
[ 6464/60000] instead of [ 6400/60000] , and so on. The Output values, e.g. loss: 2.306185  [
0/60000], read as a loss observed for zero input data which seems incorrect. This loss mentioned
here was for the first batch, which was 64 input data. The second was for 101 batch which has
6464 input data, and so on.