Use a more readable and consistent Batch variable name
There is no bug here, but this PR provides a small improvement. In N, C,
H, W variables, C=Channel, H=Height, and W=Width, which is fine, but
N=Batch (or batch size or images in a batch) is not intuitive specially
for new users and contributors. The use of B can be straightforwardly read
as a Batch (or batch size) in the context of tutorial.

I also noticed that in another doc
https://pytorch.org/vision/stable/transforms.html#torchvision.transforms.ToTensor
these variable are referenced as B, C, H, W, which is easy to understand.
So, these PR changes also make the usage of the variables consistent.