Update nn_tutorial.py
Fix to #1303 "add pyplot.show() in beginner tutorial."

Comments to issue suggested manually commenting out pyplot.show for users not using colab. 

Instead, suggest change that imports colab and does nothing if user is running notebook on colab, otherwise it runs pyplot.show() (such as when running on a local machine). 

This way, manual editing of notebook is not required, colab users are undisturbed and local users get pyplot.show.