Remove improper src_mask from encoder tutorial
The tutorial is using a transformer encoder and the mask used was for
masking a decoder which is not part of the tutorial.

The mask is removed. Some variable names are changed to better reflect
the purpose of the variable. Also, some unused imports are removed.

Fixes https://github.com/pytorch/tutorials/issues/1877

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>