Update intro.py
In the "How to Use this Guide" section, if a person is "not" familiar, then the first sentence provides reference to a page for reading about other deep learning frameworks but there is a typo  in this first sentence. I am submitting this request to correct. Please find the sentence before and after correction for reference below.

Before => If you’re familiar with other deep learning frameworks, check out the 0. Quickstart first to quickly familiarize yourself with PyTorch’s API.

After => If you’re not familiar with other deep learning frameworks, check out the 0. Quickstart first to quickly familiarize yourself with PyTorch’s API.