Add new Python and C++/CUDA Custom Op tutorials
I want to land this before PyTorch 2.4 (so we can link to these in
PyTorch's nightly documentation) and then have a follow-up PR for 2.4
that actually runs the scripts (so that they can generate outputs).

https://github.com/pytorch/pytorch/issues/127443 to remind myself of the
above.

NB: These two tutorials replace all of the existing custom ops (and cpp
extensions) tutorials:
- advanced/cpp_extension
- advanced/torch_script_custom_ops
- advanced/torch_script_custom_classes
- advanced/dispatcher

In a follow-up PR we will add warnings to all of those tutorials stating
that they are deprecated but we will preserve the text just in case
people still need them (e.g. if they are not using PyTorch 2.4).

Test Plan:
- I tested these locally.