Add batchnorm example and add thumbnail
Instead of using resnet as the example for visualization of the gradients,
I decided to use a simple fully-connected network with and without batchnorm.

It is a contrived model, but the importance is on illustration of the
gradients, not so much on which model to apply it for. I also wanted
the positive effect of batch normalization to be clearly shown, and this
was not the case with PyTorch's base resnet model.