feat: add a recipe on regional aot. (#3543)
I don't see an issue for this tutorial. I discussed this with @angelayi
and we agreed that having a recipe like is beneficial to the AoT
workflow.

I ran the Python script introduced in this PR on a single RTX 4090, and
I got:

```
Full model compilation time = 5.91 seconds
Regional compilation time = 2.54 seconds
```

cc @avikchaudhuri @gmagogsfm @zhxchen17 @tugsbayasgalan @angelayi @suo
@ydwu4

---------

Co-authored-by: Angela Yi <angelayi@meta.com>
Co-authored-by: Svetlana Karslioglu <svekars@meta.com>