Switch from sin(x) to exp(x) as the target function in polynomial_autograd  (#3597)
Fixes #3596

This PR branch uses 'exp(x)' instead of 'sin(x)' as the 'y' (target)
variable. The reason for this is faster and clearer convergence, since
exp(x) is well approximated by a Taylor expansion around the origin.
Also, all the derivatives are exp(0)=1, so the polynomial coefficients
are trivially '1/n! '(for the 'n-th' term).

I am also improving the information output to the user inside the
optimization loop.

Finally, I am making the top doc string a raw string, to avoid a python3
warning about it.


cc @albanD @jbschlosser

---------

Co-authored-by: Svetlana Karslioglu <svekars@meta.com>